lambda/requirements.txt
strands_agents
strands-agents-tools
aws_lambda_typing
fastapi
uvicorn
